event-webservice/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.29.0

pydantic==1.10.14
python-dotenv==1.0.1

peewee==3.17.1
pymysql==1.1.0

pymongo==4.6.0

grpcio
grpcio-tools
